correcoes da primeira sprint

+++ frontend/src/pages/login/index.js
@@ -11,7 +11,14 @@ function Login() {
   const [password, setPassword] = useState("");
   const navigate = useNavigate();
   
-  localStorage.clear();
+  console.log(localStorage.getItem("historico"))
+  if (localStorage.getItem("historico")) {
+    localStorage.clear();
+  }
+   
+  if (localStorage.getItem("role")) {
+    localStorage.clear();
+  }
 
   const handleSubmit = async (event) => {
     event.preventDefault();
@@ -20,9 +27,6 @@ function Login() {
         email: email,
         password: password,
       });
-
-      console.log(loginResponse);
-      
 
       if (loginResponse.status === 200) {
         const token = loginResponse.data;
